implemented core of tracking package. corrected DatabaseModule.kt

+++ data/src/main/java/com/pseteamtwo/allways/settings/TrackingRegularity.kt
@@ -1,6 +1,6 @@
 package com.pseteamtwo.allways.settings
 
-enum class TrackingRegularity(val regularity: Int) {
+enum class TrackingRegularity(val regularity: Long) {
     NEVER(0),
     RARELY(60),
     MEDIUM(30),
